# backend/fastapi/openapi-fastapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Workma!te AI API",
    "description": "회의체 운영 AI Agent 서비스의 **AI · 실시간 · 지식그래프** 백엔드 (FastAPI).\n\n- **인증**: Spring이 발급한 JWT access token을 `Authorize`(HTTPBearer)에 입력하세요. (`JWT_SECRET`을 Spring과 공유, HS256)\n- **스트리밍**: 다수 엔드포인트가 SSE(`text/event-stream`)로 토큰 단위 응답합니다.\n- **데이터 스코프**: AI 도구는 사용자의 소속 회의체로 접근을 제한합니다(IDOR 이중 방어).\n- 인증·도메인 CRUD는 별도 Spring Boot 서버(`/api/v1`)가 담당합니다.",
    "contact": {
      "name": "Team No.9"
    },
    "license": {
      "name": "Internal"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/ai/companies/rename": {
      "patch": {
        "tags": [
          "ai"
        ],
        "summary": "회사명 변경",
        "description": "회사명 변경 — SYSTEM_ADMIN 또는 해당 회사 COMPANY_ADMIN만.\n\n그래프 회사 노드는 이름으로 식별되므로 old_name으로 회사를 찾는다.\ncompanies.name 한 행만 바꾸면 users.company_name(파생 property)이 전부 따라온다.",
        "operationId": "rename_company_endpoint_api_ai_companies_rename_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/ai/departments/rename": {
      "patch": {
        "tags": [
          "ai"
        ],
        "summary": "부서명 변경",
        "description": "부서명 변경 — SYSTEM_ADMIN 또는 해당 회사 COMPANY_ADMIN만.\n\n부서는 별도 테이블 없이 users.department(문자열)로만 존재하므로\nold_name과 일치하는 users 행의 department를 일괄 변경한다.\ncompany_name이 주어지면 그 회사 소속으로 한정해 타 회사 동명 부서를 보호한다.",
        "operationId": "rename_department_endpoint_api_ai_departments_rename_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/ai/company/members": {
      "get": {
        "tags": [
          "ai"
        ],
        "summary": "회사 구성원 목록 조회",
        "description": "회사 관리 탭 데이터. 역할별 가시 범위(MT-3):\n- SYSTEM_ADMIN: 전체 사용자 (회사 경계 무시)\n- COMPANY_ADMIN: 자기 회사 소속 사용자만\n- 일반 사용자: 본인이 속한 회의체의 구성원만",
        "operationId": "company_members_api_ai_company_members_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/ai/users/{user_id}": {
      "delete": {
        "tags": [
          "ai"
        ],
        "summary": "사용자 계정 삭제",
        "description": "사용자 계정 삭제(소프트 삭제) — SYSTEM_ADMIN 또는 같은 회사 COMPANY_ADMIN만 (MT-6).\n\n하드 삭제는 작성한 회의록·보고서·감사 로그 등 users(id)를 RESTRICT로 참조하는\n행이 하나라도 있으면 불가능했다(사실상 활동한 모든 사용자 삭제 불가). 대신 행은\n남기고 is_active=False로 비활성화한다.\n- 표시용 이름(name)은 보존 → 과거 작성물의 작성자 표기 유지\n- 로그인 자격(email)은 익명화 → ① 재로그인 차단 ② 같은 이메일로 재가입 허용",
        "operationId": "ai_delete_user_api_ai_users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "ai"
        ],
        "summary": "사용자 정보 수정",
        "description": "사용자 정보(이름·이메일·회사·부서·직책)를 수정하고 Neo4j에 동기화합니다.\n\n권한: 본인 또는 SYSTEM_ADMIN, 같은 회사 COMPANY_ADMIN만 가능(MT-1).\n이메일(로그인 자격) 변경은 관리자만 허용 — 본인 자가수정으로는 불가.",
        "operationId": "ai_update_user_api_ai_users__user_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}